+++ learning/plr.h
@@ -76,7 +76,7 @@ private:
 
 public:
     PLR(double gamma);
-    std::vector<Segment> train(std::vector<std::pair<Slice, uint64_t> >& keys, bool file);
+    std::vector<Segment> train(std::vector<std::pair<Slice, long double> >& keys, bool file);
 //    std::vector<double> predict(std::vector<double> xx);
 //    double mae(std::vector<double> y_true, std::vector<double> y_pred);
 };
